refactor(output): Extract file offset tree functions into fileTreeOffsets.ts
fileTreeGenerate.ts exceeded the 250-line limit after adding the three
offset-annotation functions. Move treeToStringWithFileOffsets,
generateTreeStringWithFileOffsets, and generateTreeStringWithRootsAndFileOffsets
into a new src/core/file/fileTreeOffsets.ts. Export generateMultiRootSections
and sortTreeNodes from fileTreeGenerate.ts to support the new module.

Also add --show-file-offsets and output.showFileOffsets to README.md per
CONTRIBUTING.md documentation requirements.
